move the match method to the manager not the model
