+++ st2client/st2client/models/action_alias.py
@@ -13,8 +13,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from six.moves.urllib.parse import urlencode
-
 from st2client.models import core
 
 __all__ = [
@@ -29,13 +27,3 @@ class ActionAlias(core.Resource):
     _plural_display_name = 'Runners'
     _url_path = 'actionalias'
     _repr_attributes = ['name', 'pack', 'action_ref']
-
-    @core.add_auth_token_to_kwargs_from_env
-    def match(self, command, **kwargs):
-        url = '/%s/match' % self.resource.get_url_path_name()
-        query_str = urlencode({'command': command})
-        response = self.client.post(url, query_str, **kwargs)
-        if response.status_code != 201:
-            self.handle_error(response)
-        instance = self.resource.deserialize(response.json())
-        return instance
