+++ st2client/st2client/models/core.py
@@ -21,6 +21,8 @@ from functools import wraps
 import six
 
 from six.moves import urllib
+from six.moves.urllib.parse import urlencode
+
 from st2client.utils import httpclient
 
 
@@ -336,6 +338,16 @@ class ActionAliasResourceManager(ResourceManager):
         self.resource = resource
         self.debug = debug
         self.client = httpclient.HTTPClient(root=endpoint, cacert=cacert, debug=debug)
+
+    @add_auth_token_to_kwargs_from_env
+    def match(self, command, **kwargs):
+        url = '/%s/match' % self.resource.get_url_path_name()
+        query_str = urlencode({'command': command})
+        response = self.client.post(url, query_str, **kwargs)
+        if response.status_code != 201:
+            self.handle_error(response)
+        instance = self.resource.deserialize(response.json())
+        return instance
 
 
 class LiveActionResourceManager(ResourceManager):
